A Decorator for manifolds

A decorator manifold extends the functionality of a AbstractManifold in a semi-transparent way. It internally stores the AbstractManifold it extends and by default for functions defined in ManifoldsBase.jl it acts transparently in the sense that it passes all functions through to the base except those that it actually affects. For example, because the ValidationManifold affects nearly all functions, it overwrites nearly all functions, except a few like manifold_dimension.

By default, i.e. for a plain new decorator, all functions are transparent, i.e. passed down to the manifold the AbstractDecoratorManifold decorates. To implement a method for a decorator that behaves differently from the method of the same function for the internal manifold, two steps are required. Let's assume the function is called f(M, arg1, arg2), and our decorator manifold DM of type OurDecoratorManifold decorates M. Then

set decorator_transparent_dispatch(f, M::OurDecoratorManifold, args...) = Val(:intransparent)

implement f(DM::OurDecoratorManifold, arg1, arg2)

This makes it possible to extend a manifold or all manifolds with a feature or replace a feature of the original manifold.

A final technical note – if several manifolds have similar transparency rules concerning functions from the interface, the last parameter T of the AbstractDecoratorManifold{𝔽,T<:AbstractDecoratorType} can be used to dispatch on different transparency schemes.

ManifoldsBase.AbstractDecoratorManifold — Type

AbstractDecoratorManifold{𝔽,T<:AbstractDecoratorType} <: AbstractManifold{𝔽}

An AbstractDecoratorManifold indicates that to some extent a manifold subtype decorates another AbstractManifold in the sense that it either

it extends the functionality of a manifold with further features

it defines a new manifold that internally uses functions from the decorated manifold

with the main intent that several or most functions of AbstractManifold are transparently passed through to the manifold that is decorated. This way a function implemented for a decorator acts transparent on all other decorators, i.e. they just pass them through. If the decorator the function is implemented for is not among the decorators, an error is issued. By default all base manifold functions, for example exp and log are transparent for all decorators.

Transparency of functions with respect to decorators can be specified using the macros @decorator_transparent_fallback, @decorator_transparent_function and @decorator_transparent_signature.

There are currently three modes given a new AbstractDecoratorManifold M

:intransparent – this function has to be implmented for the new manifold M

:transparent – this function is transparent, in the sense that the function is invoked on the decorated M.manifold. This is the default, when introducing a function or signature.

:parent specifies that (unless implemented) for this function, the classical inheritance is issued, i.e. the function is invoked on Ms supertype.

ManifoldsBase.AbstractDecoratorType — Type

AbstractDecoratorType

Decorator types can be used to specify a basic transparency for an AbstractDecoratorManifold. This can be seen as an initial (rough) transparency pattern to start a type with.

Note that for a function f and it's mutating variant f!

The function f is set to :parent to first invoke allocation and call of f!

The mutating function f! is set to transparent

ManifoldsBase.DefaultDecoratorType — Type

DefaultDecoratorType <: AbstractDecoratorType

A default decorator type, where all new functions are transparent by default.

ManifoldsBase.@decorator_transparent_fallback — Macro

@decorator_transparent_fallback(ex)
@decorator_transparent_fallback(fallback_case = :intransparent, ex)

This macro introduces an additional implementation for a certain additional case. This can especially be used if for an already transparent function and an abstract intermediate type a change in the default is required. For implementing a concrete type, neither this nor any other trick is necessary. One just implements the function as before. Note that a decorator that is_default_decorator still dispatches to the transparent case.

:transparent states, that the function is transparently passed on to the manifold that is decorated by the AbstractDecoratorManifold M, which is determined using the function decorated_manifold.

:intransparent states that an implementation for this decorator is required, and if none of the types provides one, an error is issued. Since this macro provides such an implementation, this is the default.

:parent states, that this function passes on to the supertype instead of to the decorated manifold.

Inline definitions are not supported. The function signature however may contain keyword arguments and a where clause. It does not allow for parameters with default values.

Examples

@decorator_transparent_fallback function log!(M::AbstractGroupManifold, X, p, q)
    log!(decorated_manifold(M), X, p, Q)
end
@decorator_transparent_fallback :transparent function log!(M::AbstractGroupManifold, X, p, q)
    log!(decorated_manifold(M), X, p, Q)
end

ManifoldsBase.@decorator_transparent_function — Macro

@decorator_transparent_function(ex)
@decorator_transparent_function(fallback_case = :intransparent, ex)

Introduce the function specified by ex to act transparently with respect to AbstractDecoratorManifolds. This introduces the possibility to modify the kind of transparency the implementation is done for. This optional first argument, the Symbol within fallback_case. This macro can be used to define a function and introduce it as transparent to other decorators. Note that a decorator that is_default_decorator still dispatches to the transparent case.

The cases of transparency are

:transparent states, that the function is transparently passed on to the manifold that is decorated by the AbstractDecoratorManifold M, which is determined using the function decorated_manifold.

:intransparent states that an implementation for this decorator is required, and if none of the types provides one, an error is issued. Since this macro provides such an implementation, this is the default.

:parent states, that this function passes on to the supertype instead of to the decorated manifold. Passing is performed using the invoke function where the type of manifold is replaced by its supertype.

Innkoline-definitions are not yet covered – the function signature however may contain keyword arguments and a where clause.

Examples

@decorator_transparent_function log!(M::AbstractDecoratorManifold, X, p, q)
    log!(decorated_manifold(M), X, p, Q)
end
@decorator_transparent_function :parent log!(M::AbstractDecoratorManifold, X, p, q)
    log!(decorated_manifold(M), X, p, Q)
end

ManifoldsBase.@decorator_transparent_signature — Macro

@decorator_transparent_signature(ex)

Introduces a given function to be transparent with respect to all decorators. The function is adressed by its signature in ex.

Supports standard, keyword arguments and where clauses. Doesn't support parameters with default values. It introduces a dispatch on several transparency modes

The cases of transparency are

:transparent states, that the function is transparently passed on to the manifold that is decorated by the AbstractDecoratorManifold M, which is determined using the function decorated_manifold. This is the default.

:intransparent states that an implementation for this decorator is required, and if none of the types provides one, an error is issued.

:parent states, that this function passes on to the supertype instead of to the decorated manifold.

Inline definitions are not supported. The function signature however may contain keyword arguments and a where clause.

The dispatch kind can later still be set to something different, see decorator_transparent_dispatch

Examples:

@decorator_transparent_signature log!(M::AbstractDecoratorManifold, X, p, q)
@decorator_transparent_signature log!(M::TD, X, p, q) where {TD<:AbstractDecoratorManifold}
@decorator_transparent_signature isapprox(M::AbstractDecoratorManifold, p, q; kwargs...)

ManifoldsBase.decorated_manifold — Method

decorated_manifold(M::AbstractDecoratorManifold)

Return the manifold decorated by the decorator M. Defaults to M.manifold.

ManifoldsBase.decorator_transparent_dispatch — Method

decorator_transparent_dispatch(f, M::AbstractManifold, args...) -> Val

Given a AbstractManifold M and a function f(M,args...), indicate, whether a function is Val(:transparent) or Val(:intransparent) for the (decorated) AbstractManifold M. Another possibility is, that for M and given args... the function f should invoke Ms Val(:parent) implementation, see @decorator_transparent_function for details.

ManifoldsBase.default_decorator_dispatch — Method

default_decorator_dispatch(M) -> Val

Return whether by default to dispatch the the inner manifold of a decorator (Val(true)) or not (Val(false). For more details see is_decorator_transparent.

ManifoldsBase.is_decorator_transparent — Method

is_decorator_transparent(f, M::AbstractManifold, args...) -> Bool

Given a AbstractManifold M and a function f(M, args...), indicate, whether an AbstractDecoratorManifold acts transparently for f. This means, it just passes through down to the internally stored manifold. Transparency is only defined for decorator manifolds and by default all decorators are transparent. A function that is affected by the decorator indicates this by returning false. To change this behaviour, see decorator_transparent_dispatch.

If a decorator manifold is not in general transparent, it might still pass down for the case that a decorator is the default decorator, see is_default_decorator.

ManifoldsBase.is_default_decorator — Method

is_default_decorator(M) -> Bool

For any manifold that is a subtype of AbstractDecoratorManifold, this function indicates whether a certain manifold M acts as a default decorator.

This yields that all functions are passed through to the decorated AbstractManifold if M is indicated as default. This overwrites all is_decorator_transparent values.

This yields the following advantange: For a manifold one usually implicitly assumes for example a metric. To avoid reimplementation of this metric when introducing a second metric, the first metric can be set to be the default, i.e. its implementaion is already given by the undecorated case.

Value returned by this function is determined by default_decorator_dispatch, which returns a Val-wrapped boolean for type stability of certain functions.