The Manifold Type

Number systems

ManifoldsBase.AbstractNumbers — Type

AbstractNumbers

An abstract type to represent the number system on which a manifold is built.

This provides concrete number types for dispatch. The two most common number types are the fields RealNumbers (ℝ for short) and ComplexNumbers (ℂ).

ManifoldsBase.ComplexNumbers — Type

ComplexNumbers <: AbstractNumbers
ℂ = ComplexNumbers()

The field of complex numbers.

ManifoldsBase.QuaternionNumbers — Type

QuaternionNumbers <: AbstractNumbers
ℍ = QuaternionNumbers()

The division algebra of quaternions.

ManifoldsBase.RealNumbers — Type

RealNumbers <: AbstractNumbers
ℝ = RealNumbers()

The field of real numbers.

ManifoldsBase._unify_number_systems — Method

_unify_number_systems(𝔽s::AbstractNumbers...)

Compute a number system that includes all given number systems (as sub-systems) and is closed under addition and multiplication.

ManifoldsBase.number_system — Method

number_system(M::AbstractManifold{𝔽})

Return the number system the manifold M is based on, i.e. the parameter 𝔽.

ManifoldsBase.real_dimension — Method

real_dimension(𝔽::AbstractNumbers)

Return the real dimension $\dim_ℝ 𝔽$ of the AbstractNumbers system 𝔽. The real dimension is the dimension of a real vector space with which a number in 𝔽 can be identified. For example, ComplexNumbers have a real dimension of 2, and QuaternionNumbers have a real dimension of 4.

The main type: The AbstractManifold

The main type is the AbstractManifold. It represents the manifold per se. During the documentation we will use the Euclidean Space and the Sphere (both implemented in Manifolds.jl) as easy examples to often illustrate properties and features of this interface

ManifoldsBase.AbstractManifold — Type

AbstractManifold{F}

A manifold type. The AbstractManifold is used to dispatch to different functions on a manifold, usually as the first argument of the function. Examples are the exponential and logarithmic maps as well as more general functions that are built on them like the geodesic.

The manifold is parametrized by an AbstractNumbers to distinguish for example real (ℝ) and complex (ℂ) manifolds.

For subtypes the preferred order of parameters is: size and simple value parameters, followed by the AbstractNumbers field, followed by data type parameters, which might depend on the abstract number field type.

ManifoldsBase.AbstractManifoldPoint — Type

AbstractManifoldPoint

Type for a point on a manifold. While a AbstractManifold does not necessarily require this type, for example when it is implemented for Vectors or Matrix type elements, this type can be used either

for more complicated representations,

semantic verification, or

even dispatch for different representations of points on a manifold.

Since semantic verification and different representations usually might still only store a matrix internally, it is possible to use @manifold_element_forwards and @default_manifold_fallbacks to reduce implementation overhead.

which should store information about the manifold, for example parameters inherent to the manifold.

Points and Tangent Vectors

Points and tangent vectors do not necessarily have to be typed. Usually one can just use any type. When a manifold has multiple representations, these should be distinguished by point and vector types. Then it might be that types just encapsulate a vector value. This is taken into account by the following macros, that forward several actions just to this field. Most prominently vector operations for the tangent vectors. If there is still a default case, a macro sets this type to be equivalent to calling the manifold functions just with the types field that carries the value.

ManifoldsBase.@default_manifold_fallbacks — Macro

default_manifold_fallbacks(TM, TP, TV, pfield::Symbol, vfield::Symbol)

Introduce default fallbacks for all basic functions on manifolds, for manifold of type TM, points of type TP, tangent vectors of type TV, with forwarding to fields pfield and vfield for point and tangent vector functions, respectively.

ManifoldsBase.@manifold_element_forwards — Macro

manifold_element_forwards(T, field::Symbol)
manifold_element_forwards(T, Twhere, field::Symbol)

Introduce basic fallbacks for type T (which can be a subtype of Twhere) that represents points or vectors for a manifold. Fallbacks will work by forwarding to the field passed in field`

List of forwarded functions:

allocate,

copy,

copyto!,

number_eltype (only for values, not the type itself),

similar,

==.

ManifoldsBase.@manifold_vector_forwards — Macro

manifold_vector_forwards(T, field::Symbol)
manifold_vector_forwards(T, Twhere, field::Symbol)

Introduce basic fallbacks for type T that represents vectors from a vector bundle for a manifold. Twhere is put into where clause of each method. Fallbacks work by forwarding to field passed as field.

List of forwarded functions:

basic arithmetic (*, /, \, +, -),

all things from @manifold_element_forwards,

broadcasting support.

example

@eval @manifold_vector_forwards ValidationFibreVector{TType} TType value