A Decorator for manifolds

A decorator for a manifold is the main scheme to either to explixitly add a property to a manifold. This can be donne for all properties by using the AbstractDecoratorManifold as a supertype.

In general there are two scenarios, where this is used.

Dispatching a certain default property to another manifold.

Introducing an alternate variant of a property.

Let‘s make this more precise. For the first case imagine that a manifold $\mathcal M$ is represented by points in the embedding by default. If it is even isometrically embedded, the inner product can be inherited from the embedding. This is a case for the first approach. We provide a default coupling to another manifold to dispatch certain functions over. Which ones are passed on is regulated traits and their dispatch functions.

Assume that a manifold has two different embeddings or a default representation of its points/vectors works without an embedding and there is a second way to represent the manifold in some embedding. This is the situation for the second szenario, see EmbeddedManifold to provide a concrete “coupling” between a manifold and its embedding.

implicitly or explicitly add properties _

ManifoldsBase.AbstractDecoratorManifold — Type

AbstractDecoratorManifold{𝔽} <: AbstractManifold{𝔽}

Declare a manifold to be an abstract decorator. A manifold which is a subtype of is a decorated manifold, i.e. has

certain additional properties or

delegates certain properties to other manifolds.

Most prominently, a manifold might be an embedded manifold, i.e. points on a manifold $\mathcal M$ are represented by (some, maybe not all) points on another manifold $\mathcal N$. Depending on the type of embedding, several functions are dedicated to the embedding. For example if the embedding is isometric, then the inner does not have to be implemented for $\mathcal M$ but can be automatically implemented by deligation to $\mathcal N$.

This is modelled by the AbstractDecoratorManifold and traits. These are mapped to functions, which determine the types of transparencies.

A dault function to implement determines the generic manifold that is added (decorates the manifold), see decorated_manifold.

ManifoldsBase.decorated_manifold — Method

decorated_maniofold(M::AbstractDecoratorManifold)

return the manifold that is actually decorated. For an abstract case this is still the menifold itself, but for example for the EmbeddedManifold it returns the base manifold without its embedding.

ManifoldsBase.get_embedding — Method

get_embedding(M::AbstractDecoratorManifold)

Specify the embedding of a manifold that has abstract decorators.

ManifoldsBase.is_embedded_manifold — Method

IsEmbeddedManifold{M}
is_embedded_manifold(M::Type{<:AbstractDecoratorManifold})

Determine whether an AbstractDecoratorManifold M is an embedded manifold. To activate this for your manifold, set isembedded_manifold for your manifold type to true. Manifolds that are is_isometric_embedded_manifolds set this to true as well.

ManifoldsBase.is_embedded_submanifold — Method

IsEmbeddedSubmanifold{M}
is_embedded_submanifold(M::Type{<:AbstractDecoratorManifold})

Determine whether an AbstractDecoratorManifold M is an embedded submanifold. To activate this for your manifold, set is_embedded_submanifold for your manifold type to true.

Here, all retraction, inverse retractions and vectors transports, especially exp, log, and parallel_transport_to are passed to the embedding.

ManifoldsBase.is_isometric_embedded_manifold — Method

IsIsometricEmbeddedManifold{M}
is_isometric_embedded_manifold(M::Type{<:AbstractDecoratorManifold})

Determine whether an AbstractDecoratorManifold M is an isometrically embedded manifold. To activate this for your manifold, set is_isometric_embedded_manifold for your manifold type to true.

Here, for example inner and norm are passed to the embedding

This is automatically set to true, when we have an is_embedded_submanifold.