A Decorator for manifolds

A decorator for a manifold is the main scheme to either to explixitly add a property to a manifold. This can be donne for all properties by using the AbstractDecoratorManifold as a supertype.

In general there are two scenarios, where this is used.

Dispatching a certain default property to another manifold.

Introducing an alternate variant of a property.

Let‘s make this more precise. For the first case imagine that a manifold $\mathcal M$ is represented by points in the embedding by default. If it is even isometrically embedded, the inner product can be inherited from the embedding. This is a case for the first approach. We provide a default coupling to another manifold to dispatch certain functions over. Which ones are passed on is regulated traits and their dispatch functions.

Assume that a manifold has two different embeddings or a default representation of its points/vectors works without an embedding and there is a second way to represent the manifold in some embedding. This is the situation for the second szenario, see EmbeddedManifold to provide a concrete “coupling” between a manifold and its embedding.

implicitly or explicitly add properties _

ManifoldsBase.IsEmbeddedManifold — Type

IsEmbeddedManifold <: AbstractTrait

A trait to declare an AbstractManifold as an embedded manifold.

ManifoldsBase.IsEmbeddedSubmanifoldManifold — Type

IsEmbeddedSubmanifold{M}
is_embedded_submanifold(M::Type{<:AbstractDecoratorManifold})

Determine whether an AbstractDecoratorManifold M is an embedded submanifold. It is a special case of an IsIsometricEmbeddedManifold.

Here, additionally, all retraction, inverse retractions and vectors transports, especially exp, log, and parallel_transport_to are passed to the embedding.

ManifoldsBase.IsIsometricEmbeddedManifold — Type

IsIsometricManifoldEmbeddedManifold <: AbstractTrait

Determine whether an AbstractDecoratorManifold M is an isometrically embedded manifold. To activate this for your manifold, set is_isometric_embedded_manifold for your manifold type to true.

Here, for example inner and norm are passed to the embedding

ManifoldsBase.decorated_manifold — Method

decorated_manifold(M::AbstractDecoratorManifold)

For a manifold M that is decorated with properties (for example an embedding N) this function returns the manifold that is attached (as a decorator). Hence for the embedding example this is N.

ManifoldsBase.get_embedding — Method

get_embedding(M::AbstractDecoratorManifold)

Specify the embedding of a manifold that has abstract decorators.